reduce size

+++ src/cdef.h
@@ -70,8 +70,8 @@
 
 // to help record in hash
 struct Clue {
-	int num[14];
-	int count;
+	uint8_t num[14];
+	uint16_t count;
 	uint64_t hash;
 };
